refactor(bigtable): connection refresh into own file
