+++ google/cloud/bigtable/google_cloud_cpp_bigtable.bzl
@@ -73,6 +73,7 @@ google_cloud_cpp_bigtable_hdrs = [
     "internal/bulk_mutator.h",
     "internal/client_options_defaults.h",
     "internal/common_client.h",
+    "internal/connection_refresh_state.h",
     "internal/convert_policies.h",
     "internal/defaults.h",
     "internal/google_bytes_traits.h",
@@ -149,7 +150,7 @@ google_cloud_cpp_bigtable_srcs = [
     "internal/bigtable_stub.cc",
     "internal/bigtable_stub_factory.cc",
     "internal/bulk_mutator.cc",
-    "internal/common_client.cc",
+    "internal/connection_refresh_state.cc",
     "internal/convert_policies.cc",
     "internal/defaults.cc",
     "internal/google_bytes_traits.cc",
