+++ google/cloud/bigtable/internal/connection_refresh_state_test.cc
